+++ backend/src/main/java/org/rdfarchitect/api/controller/datasets/graphs/classes/ClassAllAssociationsRESTController.java
@@ -27,9 +27,9 @@ import org.rdfarchitect.api.controller.Response;
 import org.rdfarchitect.api.dto.association.AssociationPairDTO;
 import org.rdfarchitect.database.GraphIdentifier;
 import org.rdfarchitect.services.ExpandURIUseCase;
-import org.rdfarchitect.services.update.classes.associations.UpdateAssociationsUseCase;
-import org.rdfarchitect.services.update.classes.associations.CreateAssociationUseCase;
 import org.rdfarchitect.services.update.classes.associations.AssociationsService;
+import org.rdfarchitect.services.update.classes.associations.CreateAssociationUseCase;
+import org.rdfarchitect.services.update.classes.associations.UpdateAssociationsUseCase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
