+++ backend/src/main/java/org/rdfarchitect/api/controller/datasets/graphs/classes/associations/ClassAssociationsRESTController.java
@@ -25,8 +25,8 @@ import lombok.RequiredArgsConstructor;
 import org.rdfarchitect.api.dto.association.AssociationPairDTO;
 import org.rdfarchitect.database.GraphIdentifier;
 import org.rdfarchitect.services.ExpandURIUseCase;
-import org.rdfarchitect.services.update.classes.associations.UpdateAssociationsUseCase;
 import org.rdfarchitect.services.update.classes.associations.AssociationsService;
+import org.rdfarchitect.services.update.classes.associations.UpdateAssociationsUseCase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
