+++ backend/src/main/java/org/rdfarchitect/shacl/assessment.md
