+++ backend/src/main/java/org/rdfarchitect/shacl/property/shapegenerator/InverseCardinalityPropertyShapeFromCIMAssociationGenerator.java
@@ -15,19 +15,22 @@
  *
  */
 
-package org.rdfarchitect.shacl.generator.property.shapegenerator;
+package org.rdfarchitect.shacl.property.shapegenerator;
 
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.riot.system.PrefixEntry;
+import org.apache.jena.vocabulary.RDFS;
+import org.rdfarchitect.cim.rdf.resources.CIMS;
+import org.rdfarchitect.cim.relations.model.CIMClassUtils;
 import org.rdfarchitect.cim.relations.model.properties.CIMAssociationUtils;
 import org.rdfarchitect.cim.relations.model.properties.CIMPropertyUtils;
-import org.rdfarchitect.shacl.generator.property.CIMPropertySHACLUtils;
-import org.rdfarchitect.shacl.generator.property.shapebuilder.CardinalityPropertyShapeBuilder;
+import org.rdfarchitect.shacl.property.CIMPropertySHACLUtils;
+import org.rdfarchitect.shacl.property.shapebuilder.InverseCardinalityPropertyShapeBuilder;
 
-public class CardinalityPropertyShapeFromCIMPropertyGenerator implements PropertyShapeFromCIMPropertyGenerator {
+public class InverseCardinalityPropertyShapeFromCIMAssociationGenerator implements PropertyShapeFromCIMPropertyGenerator {
 
-    private static final String PROPERTY_GROUP_LABEL = "CardinalityGroup";
+    private static final String PROPERTY_GROUP_LABEL = "InverseCardinalityGroup";
 
     private Model ontologyModel;
 
@@ -54,29 +57,32 @@ public class CardinalityPropertyShapeFromCIMPropertyGenerator implements Propert
     }
 
     @Override
-    public Resource createPropertyShape(Resource property) {
+    public Resource createPropertyShape(Resource association) {
         if (ontologyModel == null || shaclModel == null || shaclPrefix == null) {
             throw new IllegalStateException("Models and prefix must be set before creating property shapes.");
         }
-        var propertyType = "unknown";
-        if (CIMPropertyUtils.isAttribute(property)) {
-            propertyType = "attribute";
+        if (!CIMPropertyUtils.isAssociation(association) || !shouldCreateInverseCardinalityPropertyShape(association)) {
+            return null; // This converter only creates shapes for associations that have an inverse Cardinality Constraint for an instantiable class
         }
-        if (CIMPropertyUtils.isAssociation(property)) {
-            if(!CIMAssociationUtils.isUsedAssociation(property)){
-                return null; // This converter only creates shapes for used associations
-            }
-            propertyType = "association";
-        }
-        var order = CIMPropertySHACLUtils.getOrder(ontologyModel, property.getURI());
-        var multiplicity = CIMPropertyUtils.resolveMultiplicity(property);
-        return new CardinalityPropertyShapeBuilder(shaclModel, propertyType)
+        var order = CIMPropertySHACLUtils.getOrder(ontologyModel, association.getURI());
+        var multiplicity = CIMPropertyUtils.resolveMultiplicity(association);
+        return new InverseCardinalityPropertyShapeBuilder(shaclModel)
                 .setPrefixEntry(shaclPrefix)
-                .setPropertyUri(property.getURI())
+                .setPropertyUri(association.getURI())
+                .setInversePropertyUri(association.getProperty(CIMS.inverseRoleName).getResource().getURI())
                 .setPropertyGroupUri(shaclPrefix.getUri() + PROPERTY_GROUP_LABEL)
                 .setOrder(order)
                 .setLowerBound(multiplicity.lowerBound())
                 .setUpperBound(multiplicity.upperBound())
                 .build();
+    }
+
+    private boolean shouldCreateInverseCardinalityPropertyShape(Resource association) {
+        if (!CIMAssociationUtils.isUsedAssociation(association.getProperty(CIMS.inverseRoleName).getResource())) {
+            return false;
+        }
+        var range = association.getProperty(RDFS.range).getResource();
+        return CIMClassUtils.isInstantiableClass(range) ||
+               !CIMClassUtils.findDerivingClasses(range).isEmpty();
     }
 }
