+++ backend/src/test/java/org/rdfarchitect/shacl/generate/testdata/valid1.ttl
