+++ backend/src/test/java/org/rdfarchitect/shacl/generate/testdata/valid2.ttl
