+++ backend/src/test/java/org/rdfarchitect/shacl/generate/testdata/validAndInvalidClass.ttl
