+++ backend/src/test/java/org/rdfarchitect/shacl/generate/testdata/validWithUnknownProperty.ttl
